+++ src/components/memberForm/index.ts
@@ -1,0 +1,1 @@
+export * from './MemberForm';
