+++ src/pages/books/UpdateBook.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const UpdateBook = () => {
+  return (
+    <div>UpdateBook</div>
+  )
+}
+
+export default UpdateBook
