+++ src/pages/members/MambersInfo.jsx
@@ -41,7 +41,7 @@ const MambersInfo = () => {
           </div>
         }
         <div className='flex justify-end gap-4'>
-          <Link to="#" className="button_update">Update</Link>
+          <Link to={`/members/update/${id}`} className="button_update">Update</Link>
           <Link to="#" className="button_delete">Delete</Link>
         </div>
       </aside>
