+++ src/pages/members/UpdateMember.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const UpdateMember = () => {
+  return (
+    <div>UpdateMember</div>
+  )
+}
+
+export default UpdateMember
